GH-00 : Minor refactoring and fix typo in test package name
