+++ src/main/java/com/rcube/configmap/validators/SpecValidator.java
@@ -3,15 +3,15 @@ package com.rcube.configmap.validators;
 import com.networknt.schema.ValidationMessage;
 import com.rcube.configmap.operator.ConfigMapCustomResource;
 import com.rcube.configmap.operator.ResourceStatus;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
-
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
 
 @Component
 @Slf4j
@@ -24,7 +24,7 @@ public class SpecValidator {
             final Map<String, String> schemaResource = resource.getSpec().getConfig().getSchema();
 
             if (schemaResource == null || CollectionUtils.isEmpty(schemaResource)) {
-                log.warn("Resource [{}] doesnt have a schema, it is recommended to have one", resource.getMetadata().getName());
+                log.warn("Resource [{}] does not have a schema, it is recommended to have one", resource.getMetadata().getName());
                 return;
             }
 
@@ -33,15 +33,7 @@ public class SpecValidator {
             if (CollectionUtils.isEmpty(dataResource)) {
                 throw new RuntimeException("Missing required fields in the data configuration.");
             }
-            final Map<String, Set<ValidationMessage>> schemaValidationResult = new HashMap<>();
-            dataResource.forEach((key, data) -> {
-                if (schemaResource.containsKey(key)) {
-                    final Set<ValidationMessage> validation = jsonContentValidators.validate(schemaResource.get(key), data);
-                    if (!CollectionUtils.isEmpty(validation)) {
-                        schemaValidationResult.put(key, validation);
-                    }
-                }
-            });
+            final Map<String, Set<ValidationMessage>> schemaValidationResult = getValidationFailures(dataResource, schemaResource);
             if (!CollectionUtils.isEmpty(schemaValidationResult)) {
                 throw new RuntimeException(String.format("Invalid data format %s", schemaValidationResult));
             }
@@ -50,5 +42,20 @@ public class SpecValidator {
             resource.setStatus(new ResourceStatus(ex.getMessage()));
             throw ex;
         }
+    }
+
+    @NotNull
+    private Map<String, Set<ValidationMessage>> getValidationFailures(final Map<String, String> dataResource,
+                                                                      final Map<String, String> schemaResource) {
+        final Map<String, Set<ValidationMessage>> schemaValidationResult = new HashMap<>();
+        dataResource.forEach((key, data) -> {
+            if (schemaResource.containsKey(key)) {
+                final Set<ValidationMessage> validation = jsonContentValidators.validate(schemaResource.get(key), data);
+                if (!CollectionUtils.isEmpty(validation)) {
+                    schemaValidationResult.put(key, validation);
+                }
+            }
+        });
+        return schemaValidationResult;
     }
 }
